move components to new folders
