+++ app/info/loading.tsx
